- apply CRAI suggestions
