+++ shared/presentation/src/androidUnitTest/kotlin/network/bisq/mobile/presentation/tabs/offers/OfferbookMarketPresenterSettingsPersistenceTest.kt
@@ -1,12 +1,10 @@
 package network.bisq.mobile.presentation.tabs.offers
 
 import io.mockk.every
-import io.mockk.mockk
 import io.mockk.mockkStatic
 import io.mockk.unmockkStatic
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.ExperimentalCoroutinesApi
-import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.test.StandardTestDispatcher
 import kotlinx.coroutines.test.advanceUntilIdle
 import kotlinx.coroutines.test.resetMain
@@ -14,21 +12,13 @@ import kotlinx.coroutines.test.runTest
 import kotlinx.coroutines.test.setMain
 import network.bisq.mobile.domain.data.model.MarketFilter
 import network.bisq.mobile.domain.data.model.MarketSortBy
-import network.bisq.mobile.domain.data.model.offerbook.MarketListItem
-import network.bisq.mobile.domain.data.replicated.common.currency.MarketVO
-import network.bisq.mobile.domain.data.repository.SettingsRepository
 import network.bisq.mobile.domain.data.repository.SettingsRepositoryMock
-import network.bisq.mobile.domain.service.market_price.MarketPriceServiceFacade
-import network.bisq.mobile.domain.service.offers.OffersServiceFacade
-import network.bisq.mobile.domain.service.user_profile.UserProfileServiceFacade
 import network.bisq.mobile.domain.utils.CoroutineJobsManager
-import network.bisq.mobile.presentation.common.test_utils.MainPresenterTestFactory
 import network.bisq.mobile.presentation.common.test_utils.NoopNavigationManager
-import network.bisq.mobile.presentation.common.test_utils.TestApplicationLifecycleService
+import network.bisq.mobile.presentation.common.test_utils.OfferbookMarketPresenterTestFactory
 import network.bisq.mobile.presentation.common.test_utils.TestCoroutineJobsManager
 import network.bisq.mobile.presentation.common.ui.navigation.manager.NavigationManager
 import network.bisq.mobile.presentation.common.ui.platform.getScreenWidthDp
-import network.bisq.mobile.presentation.tabs.offers.usecase.ComputeOfferbookMarketListUseCase
 import org.koin.core.context.startKoin
 import org.koin.core.context.stopKoin
 import org.koin.dsl.module
@@ -69,41 +59,11 @@ class OfferbookMarketPresenterSettingsPersistenceTest {
         }
     }
 
-    private fun createPresenter(settingsRepository: SettingsRepository): OfferbookMarketPresenter {
-        val mainPresenter =
-            MainPresenterTestFactory.create(applicationLifecycleService = TestApplicationLifecycleService())
-
-        val offersServiceFacade = mockk<OffersServiceFacade>(relaxed = true)
-        val userProfileServiceFacade = mockk<UserProfileServiceFacade>(relaxed = true)
-        every { userProfileServiceFacade.ignoredProfileIds } returns MutableStateFlow(emptySet())
-
-        val marketPriceServiceFacade =
-            object : MarketPriceServiceFacade(settingsRepository) {
-                override fun findMarketPriceItem(marketVO: MarketVO) = null
-
-                override fun findUSDMarketPriceItem() = null
-
-                override fun refreshSelectedFormattedMarketPrice() {
-                }
-
-                override fun selectMarket(marketListItem: MarketListItem): Result<Unit> = Result.success(Unit)
-            }
-
-        return OfferbookMarketPresenter(
-            mainPresenter = mainPresenter,
-            offersServiceFacade = offersServiceFacade,
-            marketPriceServiceFacade = marketPriceServiceFacade,
-            userProfileServiceFacade = userProfileServiceFacade,
-            settingsRepository = settingsRepository,
-            computeOfferbookMarketListUseCase = ComputeOfferbookMarketListUseCase(marketPriceServiceFacade),
-        )
-    }
-
     @Test
     fun `changing market sort by from UI persists to SettingsRepository`() =
         runTest(testDispatcher) {
             val settingsRepository = SettingsRepositoryMock()
-            val presenter = createPresenter(settingsRepository)
+            val presenter = OfferbookMarketPresenterTestFactory.create(settingsRepository)
 
             presenter.setSortBy(MarketSortBy.NameAZ)
             advanceUntilIdle()
@@ -115,7 +75,7 @@ class OfferbookMarketPresenterSettingsPersistenceTest {
     fun `changing market filter from UI persists to SettingsRepository`() =
         runTest(testDispatcher) {
             val settingsRepository = SettingsRepositoryMock()
-            val presenter = createPresenter(settingsRepository)
+            val presenter = OfferbookMarketPresenterTestFactory.create(settingsRepository)
 
             presenter.setFilter(MarketFilter.WithOffers)
             advanceUntilIdle()
